+++ changes.md
@@ -9,6 +9,8 @@
 - Updated `mod/questionnaire`
 - Updated `mod/subcourse`
 - Updated `blocks/completion_progress`
+- Updated `question/type/formulas`
+- Updated `course/format/remuiformat`
 - Updated Wiris plugins
 
 ### 4.1.10+ (Build: 20240503)